requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
bleach==6.0.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
dataclasses-json==0.5.7
deprecation==2.1.0
docutils==0.19
dotty-dict==1.3.1
filelock==3.10.3
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.0
greenlet==2.0.2
h11==0.14.0
hikari==2.0.0.dev118
hikari-lightbulb==2.3.3
httpcore==0.16.3
httpx==0.23.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.1.0
invoke==1.7.3
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.13.1
langchain==0.0.113
llama-index==0.4.29
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
pkginfo==1.9.6
postgrest==0.10.6
protobuf==4.22.1
psycopg2==2.9.5
pycodestyle==2.10.0
pycryptodomex==3.17
pydantic==1.10.6
Pygments==2.14.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.13.0
python-semantic-release==7.33.2
pytz==2022.7.1
pywin32-ctypes==0.2.0
PyYAML==6.0
readme-renderer==37.3
realtime==1.0.0
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.46
stanfordnlp==0.2.0
storage3==0.5.2
StrEnum==0.4.9
supabase==1.0.2
supafunc==0.2.2
sympy==1.11.1
tenacity==8.2.2
tiktoken==0.3.1
tomlkit==0.11.6
torch==2.0.0
tqdm==4.65.0
twine==3.8.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
webencodings==0.5.1
websockets==10.4
yarl==1.8.2
yt-dlp==2023.3.4
zipp==3.15.0
